requirements.txt
anyio==4.12.1
asgiref==3.9.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.9.0
django-elasticsearch-dsl==9.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==25.2
django-nine==0.2.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
elastic-transport==9.2.1
elasticsearch==9.3.0
elasticsearch-dsl==0.0.9
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
packaging==26.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
redis==7.0.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.6.3
zope.interface==8.0.1
drf-spectacular==0.27.2